refactors: Change name of dashboard and start overview page
