--- .env
@@ -1,6 +1,0 @@
-VITE_APP_VERSION=v1.3.0
-GENERATE_SOURCEMAP=false
-
-## Backend API URL
-PUBLIC_URL = https://mantisdashboard.io/free
-VITE_APP_BASE_NAME = /free
