+++ .env.example
@@ -1,0 +1,5 @@
+VITE_APP_VERSION=v1.3.0
+GENERATE_SOURCEMAP=false
+
+## Backend API URL
+VITE_APP_BASE_NAME=/app
