+++ .gitignore
@@ -106,3 +106,6 @@ dist
 
 # wincompare file
 .bak
+
+.idea
+.env
