+++ src/api/menu.js
@@ -12,7 +12,7 @@ const initialState = {
 export const endpoints = {
   key: 'api/menu',
   master: 'master',
-  dashboard: '/dashboard' // server URL
+  dashboard: '/overview' // server URL
 };
 
 export function useGetMenuMaster() {
